MVP v 0.3.6 Cart and Wallet

+++ src/main/java/ru/psuti/blarket/dto/announcement/CreateAnnouncementDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class CreateAnnouncementDTO {
     private String title;
     private String description;
-    private BigDecimal price;
+    private Double price;
     private List<String> imageUrls;
     private String address;
     private Integer quantity;
